refactor

+++ src/Models/OptionsModel.php
@@ -6,6 +6,7 @@ class OptionsModel extends Model
 
     function create( array $fields )
     {
+        $fields = $this->secureFields( $fields );
         $this->saveOptions( $fields );
 
         return $this;
@@ -13,6 +14,7 @@ class OptionsModel extends Model
 
     function update( $itemId, array $fields )
     {
+        $fields = $this->secureFields( $fields );
         $this->saveOptions( $fields );
 
         return $this;
@@ -20,20 +22,23 @@ class OptionsModel extends Model
 
     private function saveOptions( array $fields )
     {
-        foreach ($fields as $key => $value) :
+        if ( ! empty( $fields )) {
+            foreach ($fields as $key => $value) :
 
-            if (is_string( $value )) {
-                $value = trim( $value );
-            }
+                if (is_string( $value )) {
+                    $value = trim( $value );
+                }
 
-            $current_meta = get_option( $key );
+                $current_meta = get_option( $key );
 
-            if (( isset( $value ) && $value !== "" ) && $current_meta !== $value) :
-                update_option( $key, $value );
-            elseif ( ! isset( $value ) || $value === "" && ( isset( $current_meta ) || $current_meta === "" )) :
-                delete_option( $key );
-            endif;
+                if (( isset( $value ) && $value !== "" ) && $current_meta !== $value) :
+                    update_option( $key, $value );
+                elseif ( ! isset( $value ) || $value === "" && ( isset( $current_meta ) || $current_meta === "" )) :
+                    delete_option( $key );
+                endif;
 
-        endforeach;
+            endforeach;
+        }
+
     }
 }
